requirements.txt
langchain
langchain-community
fastapi
uvicorn[standard]
python-dotenv
pydantic
PyMuPDF
requests
openai
tiktoken
faiss-gpu
langchain-openai
numpy<2